third_party/WebKit/LayoutTests/editing/deleting/delete_link_at_end_of_sentence.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                '<span>this is a <a href="foo">link|</a></span>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                '<span>this is a <a href="foo">lin|</a></span>',,
            '</div>',
        ].join('')),
        'Delete command should not delete A element');
</script>
